remove LDAP, move PERA, rename legacy module package names from generic. (legacy-caching-rest-client & legacy-cxf-cas)
