--- java-ldap/pom.xml
@@ -1,48 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>fi.vm.sade.java-utils</groupId>
-    <artifactId>java-utils</artifactId>
-    <version>0.2.2-SNAPSHOT</version>
-    <relativePath>..</relativePath>
-  </parent>
-  <artifactId>java-ldap</artifactId>
-  <version>0.1.1-SNAPSHOT</version>
-  <packaging>jar</packaging>
-  <properties>
-    <spring.security.version>4.2.11.RELEASE</spring.security.version>
-    <spring.version>5.1.5.RELEASE</spring.version>
-  </properties>
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.springframework</groupId>
-        <artifactId>spring-core</artifactId>
-        <version>${spring.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>commons-codec</groupId>
-        <artifactId>commons-codec</artifactId>
-        <version>1.11</version>
-      </dependency>
-      <dependency>
-        <groupId>org.slf4j</groupId>
-        <artifactId>slf4j-api</artifactId>
-        <version>1.7.25</version>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-  <dependencies>
-    <dependency>
-      <groupId>fi.vm.sade.java-utils</groupId>
-      <artifactId>opintopolku-spring-security</artifactId>
-      <version>0.2.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.security</groupId>
-      <artifactId>spring-security-ldap</artifactId>
-      <version>${spring.security.version}</version>
-    </dependency>
-  </dependencies>
-</project>
