--- opintopolku-spring-security/src/main/java/fi/vm/sade/generic/common/auth/SadeUserDetailsWrapper.java
@@ -1,89 +1,0 @@
-package fi.vm.sade.generic.common.auth;
-/*
- *
- * Copyright (c) 2012 The Finnish Board of Education - Opetushallitus
- *
- * This program is free software:  Licensed under the EUPL, Version 1.1 or - as
- * soon as they will be approved by the European Commission - subsequent versions
- * of the EUPL (the "Licence");
- *
- * You may not use this work except in compliance with the Licence.
- * You may obtain a copy of the Licence at: http://www.osor.eu/eupl/
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * European Union Public Licence for more details.
- */
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-public class SadeUserDetailsWrapper implements UserDetails  {
-
-    private UserDetails details;
-    private String lang;
-
-    @Override
-    public String toString() {
-        return getUsername();
-    }
-
-    public SadeUserDetailsWrapper(UserDetails details) {
-        this.details = details;
-    }
-
-    public SadeUserDetailsWrapper(UserDetails details, String lang) {
-        this.details = details;
-        this.lang = lang;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return details.getAuthorities();
-    }
-
-    @Override
-    public String getPassword() {
-        return details.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return details.getUsername();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return details.isAccountNonExpired();
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return details.isAccountNonLocked();
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return details.isCredentialsNonExpired();
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return details.isEnabled();
-    }
-
-    public String getLang() {
-        return lang;
-    }
-
-    public void setLang(String lang) {
-        this.lang = lang;
-    }
-
-    public UserDetails getDetails() {
-        return details;
-    }
-}
